undo libgd-version experiment [skip ci]
'apt list' only showed the one version.  I am sure there must be a way to grab older versions from some non-default repo, but I don't know enough about the apt ecosystem to be able to do that.  It was worth the quick try, but not worth any more investigation for now.